+++ examples/clickhouse_integrations/credentials/project.tf
@@ -1,0 +1,3 @@
+resource "aiven_project" "clickhouse_managed_credentials" {
+  project = var.aiven_project_name
+}