+++ examples/clickhouse_integrations/credentials/provider.tf
@@ -1,0 +1,15 @@
+# Initialize the provider
+
+terraform {
+  required_version = ">=0.13"
+  required_providers {
+    aiven = {
+      source  = "aiven/aiven"
+      version = ">=4.25.0, <5.0.0"
+    }
+  }
+}
+
+provider "aiven" {
+  api_token = var.aiven_api_token
+}